// src/types/models/gprProfile.ts
export interface GprProfile {
  rectangleLineNumber: Number;
  profileType: String;
  longitudinalProfileNumber: Number;
  traversalProfileNumber: Number;
  distance: Number;
  spacing: Number;
  numberOfProfile: Number;
  startingVertexX: Number;
  startingVertexY: Number;
  startingVertexZ: Number;
  endVertexX: Number;
  endVertexY: Number;
  endVertexZ: Number;
  frequency: Number;
  filname: String;
}